Logica de datos y estados centralizados en hooks
