+++ apps/client/src/components/ItemCategory/CategoryTable.tsx
@@ -1,10 +1,4 @@
-import { useState, useEffect } from "react";
-import { categoryService } from "../../services/categoryService";
-import { CategoriaDTO,  CategoriaCrearDTO, CategoriaActualizarDTO } from "../../types/categoryTypes";
-import { itemService } from "../../services/itemService";
-import { ItemRowDTO } from '../../types/itemTypes'
-import axios from "axios";
-
+import { useCategoryLogic } from "../../hooks/useCategoryLogic";
 
 import Toast from "../UI/Toast";
 import Search  from "../ItemCategory/Search";
@@ -15,154 +9,21 @@ import { Edit2, X, Check, MoreHorizontal, Plus } from "lucide-react";
 
 const CategoryTable = () => {
 
-
-    //toast
-    const [toastMessage, setToastMessage] = useState<string | null>(null);
-    const [toastType, setToastType] = useState<"success" | "error" | "info">("success");
-
-    //categoria
-    const [categories, setCategories] = useState<CategoriaDTO[]>([]);
-    const [filteredCategories, setFilteredCategories] = useState<CategoriaDTO[]>([]);
-
-    //editar categoria
-    const [isEditing, setIsEditing] = useState(false);
-    const [categoryToEdit, setCategoryToEdit] = useState<CategoriaDTO | null>(null);
-
-    //crear categoria
-    const [isCreating, setIsCreating] = useState(false);
-
-
-    //items
-    const [items, setItems] = useState<ItemRowDTO[]>([]);
-
-    //listar categorias
-    const fetchCategories = async () => {
-    const data = await categoryService.getAll();
-    setCategories(data);
-    setFilteredCategories(data);
-    };
-
-    useEffect(() => {
-    fetchCategories();
-    }, []);
-
-
-
-
-    //crear
-    const handleCreateCategory = async (newCategory: CategoriaCrearDTO) => {
-    try {
-        await categoryService.create(newCategory);
-
-        const refreshed = await categoryService.getAll();
-        setCategories(refreshed);
-        setFilteredCategories(refreshed);
-        setIsCreating(false); // Cierra el modal si aplica
-
-        setToastType("success");
-        setToastMessage("Categoría creada exitosamente.");
-    } catch (error) {
-        let msg = "Error inesperado al crear la categoría";
-
-        if (axios.isAxiosError(error)) {
-        if (error.response?.status === 401) {
-            msg = "No autorizado. Iniciá sesión nuevamente.";
-        } else {
-            const rawMsg = error.response?.data?.message;
-            msg = Array.isArray(rawMsg) ? rawMsg.join(" - ") : rawMsg || msg;
-        }
-        }
-
-        setToastType("error");
-        setToastMessage(msg);
-    }
-    };
-
-
-
-    //editar
-    const openEditCategory = (category: CategoriaDTO) => {
-    setCategoryToEdit(category);
-    setIsEditing(true);
-    };
-
-    const handleSaveCategory = async (updatedCategory: CategoriaActualizarDTO) => {
-        if (
-            updatedCategory.nombre.trim() === categoryToEdit?.nombre.trim() &&
-            updatedCategory.descripcion.trim() === categoryToEdit?.descripcion.trim()
-        ) {
-            setToastType("info");
-            setToastMessage("No hay cambios para guardar.");
-            return;
-        }
-
-    try {
-        await categoryService.update(categoryToEdit!.id_categoria, updatedCategory);
-
-        if (!updatedCategory.nombre.trim() || !updatedCategory.descripcion.trim()) {
-        setToastType("error");
-        setToastMessage("Todos los campos son obligatorios.");
-        return;
-        }
-
-        const refreshed = await categoryService.getAll();
-        setCategories(refreshed);
-        setFilteredCategories(refreshed);
-
-
-        setIsEditing(false);
-        setToastType("success");
-        setToastMessage("Categoría modificada exitosamente.");
-    } catch (error) {
-        setToastType("error");
-        setToastMessage("Error al modificar la categoría.");
-    }
-    };
-
-
-    //conteo items
-    useEffect(() => {
-    itemService.listarItems()
-        .then((res) => setItems(res.data))
-        .catch((err) => console.error("Error al listar items:", err));
-    }, []);
-
-    //setear activo/inactivo
-   const toggleEstadoCategory = async (category: CategoriaDTO) => {
-    try {
-        if (category.estado === 1) {
-        await categoryService.disable(category.id_categoria);
-        } else {
-        await categoryService.enable(category.id_categoria);
-        }
-
-        const newCategories = await categoryService.getAll();
-        setCategories(newCategories);
-        setFilteredCategories(newCategories);
-
-
-        setToastType("success");
-        setToastMessage(
-        category.estado === 1
-            ? "Categoría desactivada correctamente"
-            : "Categoría activada correctamente"
-        );
-    } catch (error) {
-        setToastType("error");
-        setToastMessage("Error al actualizar el estado de la categoría");
-    }
-    };
-
-
-    //busqueda
-    const handleSearch = (term: string) => {
-    const lower = term.toLowerCase();
-    const filtered = categories.filter((category) =>
-        category.nombre.toLowerCase().includes(lower) ||
-        category.descripcion.toLowerCase().includes(lower)
-    );
-    setFilteredCategories(filtered);
-    };
+    const {
+    toastMessage, setToastMessage,
+    toastType, setToastType,
+    categories,
+    filteredCategories, setFilteredCategories,
+    isEditing, setIsEditing,
+    isCreating, setIsCreating,
+    categoryToEdit, setCategoryToEdit,
+    items,
+    handleCreateCategory,
+    openEditCategory,
+    handleSaveCategory,
+    toggleEstadoCategory,
+    handleSearch
+    } = useCategoryLogic();
 
 
     return (
