+++ apps/client/src/hooks/usePermisos.ts
@@ -1,0 +1,18 @@
+//implementa permisos definidos en permisos/roles
+import { useAuth } from "../context/AuthContext";
+import { permisosPorRol } from "../permisos/roles";
+
+export function usePermisos() {
+  const { userType } = useAuth();
+
+  const tienePermiso = (permiso: string) => {
+    const rol = userType;
+    if (!rol) return false;
+
+    if (rol === "admin") return true;
+
+    return permisosPorRol[rol]?.includes(permiso) ?? false;
+  };
+
+  return { tienePermiso };
+}
